requirements.txt
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
executing==2.0.1
fake-useragent==1.4.0
fastjsonschema==2.19.1
filelock==3.13.1
fqdn==1.5.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.41.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-documentai==2.23.0
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
howdoi==2.0.20
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
keep==2.10.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.26.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
Protego==0.3.0
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyGithub==2.2.0
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
referencing==0.33.0
requests==2.31.0
requests-file==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
Scrapy==2.11.0
selenium==4.17.2
Send2Trash==1.8.2
service-identity==24.1.0
shapely==2.0.2
simplekml==1.3.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
terminaltables==3.1.10
tinycss2==1.2.1
tldextract==5.1.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.0
w3lib==2.1.2
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
zope.interface==6.1
